+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: mannouao <mannouao@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 12:06:13 by mannouao          #+#    #+#             */
-/*   Updated: 2022/03/11 08:28:10 by mannouao         ###   ########.fr       */
+/*   Updated: 2022/03/12 09:52:29 by mannouao         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i] != '\0')
